fix auth in delete

+++ backend/routers/anime.py
@@ -128,10 +128,13 @@ async def anime_edit_admin(request: Request, anime_id: int, current_user: user_s
 # delete anime info from DB
 @router.delete("/{anime_id}")
 @limiter.limit("2/minute")
-async def delete_anime(request: Request, anime_id: int, db: AsyncSession = Depends(get_db)):
-    anime = await anime_crud.delete_anime(db=db, anime_id=anime_id)
-    if anime is None:
-        raise HTTPException(status_code=404, detail="Anime not found")
+async def delete_anime(request: Request, anime_id: int, current_user: user_schema.CurrentUserResponse = Depends(user_router.get_current_user_required), db: AsyncSession = Depends(get_db)):
+    if current_user.user_attribute_name != "admin":
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="管理者権限が必要です")
+    else:
+        anime = await anime_crud.delete_anime(db=db, anime_id=anime_id)
+        if anime is None:
+            raise HTTPException(status_code=404, detail="Anime not found")
     return {"message": "Anime deleted successfully"}
 
 # delete request anime from DB
